tests/submission/W3C/domhtml/HTMLSelectElement08.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>NIST DOM HTML Test - SELECT</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/domtests-common.js"></script>
<link href="http://www.w3.org/TR/html5/the-select-element.html#dom-select-options" rel="help">
<link rel="author" title="NIST">
</head>
<body>
<form action="./files/getData.pl" id="form1" method="post">
<p>
<select dir="ltr" id="selectId" multiple="multiple" name="select1" size="1" tabindex="7">
<option selected="selected" value="EMP1">EMP10001</option>
<option>EMP10002</option>
<option>EMP10003</option>
<option>EMP10004</option>
<option>EMP10005</option>
</select>
</p>
</form>
<p>
<select name="select2">
<option>EMP20001</option>
<option>EMP20002</option>
<option>EMP20003</option>
<option>EMP20004</option>
<option>EMP20005</option>
</select>
</p>
<p>
<select disabled="disabled" name="select3" tabindex="1">
<option>EMP30001</option>
<option>EMP30002</option>
<option>EMP30003</option>
<option>EMP30004</option>
<option>EMP30005</option>
</select>
</p>








<script>
var domtest = async_test("The options attribute returns a collection of OPTION elements contained     by this element.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var optionsnodeList;
      var testNode;
      var vareas;
      var doc;
      var optionName;
      var voption;
      var result = new Array();

      expectedOptions = new Array();
      expectedOptions[0] = "option";
      expectedOptions[1] = "option";
      expectedOptions[2] = "option";
      expectedOptions[3] = "option";
      expectedOptions[4] = "option";

      doc = window.document;
      nodeList = doc.getElementsByTagName("select");
      assert_equals(nodeList.length,3);
testNode = nodeList.item(0);
      optionsnodeList = testNode.options;

      for(var indexidp458064 = 0;indexidp458064 < optionsnodeList.length; indexidp458064++) {
      voption = optionsnodeList.item(indexidp458064);
      optionName = voption.nodeName;

      result[result.length] = optionName;

	}
   assert_equals_list_autocase("element", result,expectedOptions);
       
    domtest.done();
}));

</script>
<p>The options attribute returns a collection of OPTION elements contained
    by this element.

    Retrieve the options attribute from the first SELECT element and
    examine the items of the returned collection.</p>
<div id="log"></div>
</body></html>
